+++ src/App.js
@@ -3,9 +3,9 @@ import Header from './common/header';
 import Footer from './common/footer';
 import styled from 'styled-components';
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import HomePage from './pages/homePage';
-import ProductsPage from './pages/productsPage';
-import { NotFound } from './pages/notFoundPage';
+import HomePage from './pages/Home';
+import ProductsPage from './pages/Products';
+import { NotFound } from './pages/NotFoundPage';
 
 const StyledWrapper = styled.div`
   position: relative;
